implemented vector of vector architecture for rust strings and appears to be working. However, list scopes is still not working as intended

+++ PravegaCSharp/CSharpWrapperSource/Byte.cs
@@ -53,7 +53,7 @@ namespace Pravega.Byte
         internal static extern ulong ByteWriterWrite(
             IntPtr byteWriterPointer,
             IntPtr bufferPointer,
-            uint bufferSize,
+            ulong bufferSize,
             ulong key,
             [MarshalAs(UnmanagedType.FunctionPtr)] rustCallbackU64Invoke callback);
         
@@ -280,7 +280,7 @@ namespace Pravega.Byte
 
             // Split the list into local variables. 
             byte[] bufferArray = buffer.ToArray();
-            uint bufferSize = (uint)bufferArray.Length;
+            ulong bufferSize = (ulong)bufferArray.Length;
 
             // Marshal the array to unmanaged memory.
             IntPtr unmanagedBufferArray = Marshal.AllocCoTaskMem(Marshal.SizeOf(typeof(byte))
